feat: implement normalization function for possibly quoted environment variable values
